feat: esm.mjs done
