+++ src/modules/esm.mjs
@@ -1,0 +1,45 @@
+import { readFile } from "fs/promises";
+
+import { fileURLToPath } from "url";
+import { dirname, sep, join } from "path";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+import { release, version } from "node:os";
+import { createServer } from "http";
+import "./files/c.js";
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+  unknownObject = JSON.parse(
+    await readFile(join(__dirname, "files", "a.json"))
+  );
+} else {
+  unknownObject = JSON.parse(
+    await readFile(join(__dirname, "files", "b.json"))
+  );
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServer((_, res) => {
+  res.end("Request accepted");
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log("To terminate it, use Ctrl+C combination");
+});
